fix progress bar
