+++ src/pages/BasicQuestionsList.tsx
@@ -17,16 +17,16 @@ export type actualMeaning = Record<number, string>
 //QUESTIONS
 const QUESTION_1: BasicQuestionType = {instruction: "Are you an extroverted person?", answers:["Yes", "Mayhaps", "No"], id: 0}
 const QUESTION_2: BasicQuestionType = {instruction: "Are people are of value to you?", answers:["Yes", "Mayhaps", "No"], id: 1}
-const QUESTION_3: BasicQuestionType = {instruction: "Do you want to make the world a better place?", answers:["Yes", "No"], id: 2}
+const QUESTION_3: BasicQuestionType = {instruction: "Do you want to make the world a better place?", answers:["Yes", "Mayhaps", "No"], id: 2}
 const QUESTION_4: BasicQuestionType = {instruction: "Can you work in a team?", answers:["Yes", "Mayhaps", "No"], id: 3}
 const QUESTION_5: BasicQuestionType = {instruction: "Are you your own boss?", answers:["Yes", "Mayhaps", "No"], id: 4}
 const QUESTION_6: BasicQuestionType = {instruction: "Do you want to fix and make the world more perfect?", answers:["Yes", "Mayhaps", "No"], id: 5}
 const QUESTION_7: BasicQuestionType = {instruction: "Are you good with technology?", answers:["Yes", "Mayhaps", "No"], id: 6}
 const QUESTION_8: BasicQuestionType = {instruction: "Is there no cost too great?", answers:["Yes", "Mayhaps", "No"], id: 7}
-const QUESTION_9: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 8}
-const QUESTION_10: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 9}
-const QUESTION_11: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 10}
-const QUESTION_12: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 11}
+const QUESTION_9: BasicQuestionType = {instruction: "How much work experience do you have?", answers:["None", "A few months", "A year", "Years"], id: 8}
+const QUESTION_10: BasicQuestionType = {instruction: "Do you know what you are doing?", answers:["No clue", "A grasp", "Somewhat", "My actions are utterly unclouded"], id: 9}
+const QUESTION_11: BasicQuestionType = {instruction: "How much autonomy do you desire?", answers:["Give me orders and I follow", "Some guidance" ,"I'll do my own thing", "I give the orders"], id: 10}
+const QUESTION_12: BasicQuestionType = {instruction: "What is your favorite subject?", answers:["Math", "Science", "English", "PE"], id: 11}
 const QUESTION_13: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 12}
 const QUESTION_14: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 13}
 const QUESTION_15: BasicQuestionType = {instruction: "This is a plaeholder", answers:["placeholder 1", "placeholder 2"], id: 14}
