+++ src/pages/BasicQuestionsPage.css
@@ -18,6 +18,13 @@ justify-content: flex-start;
 color: #61dafb;
 }
  
+.Basic-Question-Row{
+    min-height: 12px;
+}
+
+.Basic-Question{
+    padding: 12px;
+}
 
 footer {
 text-align: center;
